+++ CHANGELOG.md
@@ -1,12 +1,30 @@
-## [1.0.2](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.1...v1.0.2) (2021-12-14)
+## [1.0.5](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.4...v1.0.5) (2021-12-18)
 
 
 ### Bug Fixes
 
-* Paris Rivoli store added ([#74](https://github.com/Ephigenia/ikea-availability-checker/issues/74)) ([8195efa](https://github.com/Ephigenia/ikea-availability-checker/commit/8195efa8f12493e7eb08e850c72bb97dd279ad0f))
-* **stores:** adds Queens, NY store ([#75](https://github.com/Ephigenia/ikea-availability-checker/issues/75)) ([431a788](https://github.com/Ephigenia/ikea-availability-checker/commit/431a78865daa58f2fb008d42429acfd4f7aeebc9))
+* **iows:** getStoreProductAvailability accepts number productIds ([7fea4dc](https://github.com/Ephigenia/ikea-availability-checker/commit/7fea4dcad92cf28dc32986ad0ffc7f091048c5b6))
+* **iows:** normalization of productCode trim after replace ([ed0415d](https://github.com/Ephigenia/ikea-availability-checker/commit/ed0415dabe4e49557c6c3d9f0729f348eb76320a))
+* **iows:** normalize buCode ([c9ce169](https://github.com/Ephigenia/ikea-availability-checker/commit/c9ce16908f5d5ffd1b4fad9b7a396586d51c2f6f))
+* **iows:** throw error on non-numeric buCode or productCode ([254616e](https://github.com/Ephigenia/ikea-availability-checker/commit/254616e20a70bf23010bc56a2cc46acfd07756c0))
 
-## [1.0.1](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.0...v1.0.1) (2021-12-14)
+## [1.0.4](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.3...v1.0.4) (2021-12-17)
+
+
+### Bug Fixes
+
+* adds all australian ikea stores ([#81](https://github.com/Ephigenia/ikea-availability-checker/issues/81)) ([a04df08](https://github.com/Ephigenia/ikea-availability-checker/commit/a04df08a6185f7281de25fc80a726e54e59ab4da))
+* **iows:** throw a IOWS2DeprecationError on deprecation header ([#80](https://github.com/Ephigenia/ikea-availability-checker/issues/80)) ([270c3a5](https://github.com/Ephigenia/ikea-availability-checker/commit/270c3a5091fdd2f8aafb2220a7e73a8145762b95))
+
+## [1.0.3](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.2...v1.0.3) (2021-12-16)
+
+
+### Bug Fixes
+
+* adds check-pr-title workflow ([d8d5d19](https://github.com/Ephigenia/ikea-availability-checker/commit/d8d5d195a20363e92a93b1776df3637c8ddba361))
+
+## [1.0.2](https://github.com/Ephigenia/ikea-availability-checker/compare/v1.0.1...v1.0.2) (2021-12-16)
+
 
 ### Bug Fixes